changed the types of a couple variables so some lines are less verbose

+++ src/input.rs
@@ -31,9 +31,9 @@ pub fn handle_keybind() {
 }
 
 pub fn handle_char_inp(editor_data: &mut util::EditorDataStruct, character: char) {
-    let mut curfile = &mut editor_data.files[editor_data.current_file_index as usize];
-    let mut curline = &mut curfile.file_lines[curfile.cursor_y as usize];
-    curline.line_text.insert((curfile.cursor_x) as usize, character);
+    let mut curfile = &mut editor_data.files[editor_data.current_file_index];
+    let mut curline = &mut curfile.file_lines[curfile.cursor_y];
+    curline.line_text.insert(curfile.cursor_x, character);
 
     curfile.cursor_x += 1;
     curline.line_length += 1;
@@ -41,14 +41,14 @@ pub fn handle_char_inp(editor_data: &mut util::EditorDataStruct, character: char
 }
 
 pub fn handle_remove(editor_data: &mut util::EditorDataStruct, isDel: bool) {
-    let mut file_data = &mut editor_data.files[editor_data.current_file_index as usize];
-    let mut curline = &mut file_data.file_lines[file_data.cursor_y as usize];
+    let mut file_data = &mut editor_data.files[editor_data.current_file_index];
+    let mut curline = &mut file_data.file_lines[file_data.cursor_y];
 
     if isDel {
-        curline.line_text.remove(file_data.cursor_x as usize);
+        curline.line_text.remove(file_data.cursor_x);
     } else {
 
-        if curline.line_length == editor_data.settings.line_endings as usize {
+        if curline.line_length == editor_data.settings.line_endings {
             file_data.file_lines.remove(file_data.cursor_y);
             file_data.cursor_y -= 1;
             curline = &mut file_data.file_lines[file_data.cursor_y];
@@ -56,19 +56,19 @@ pub fn handle_remove(editor_data: &mut util::EditorDataStruct, isDel: bool) {
             return;
         }
         if file_data.cursor_x == 0 && curline.line_length > editor_data.settings.line_endings {
-            let cursline = file_data.file_lines[file_data.cursor_y as usize].line_text.clone();
-            file_data.file_lines.remove(file_data.cursor_y as usize);
+            let cursline = file_data.file_lines[file_data.cursor_y].line_text.clone();
+            file_data.file_lines.remove(file_data.cursor_y);
             file_data.cursor_y -= 1;
-            file_data.file_lines[file_data.cursor_y as usize].line_text.pop();
-            file_data.file_lines[file_data.cursor_y as usize].line_text.pop();
-            file_data.file_lines[file_data.cursor_y as usize].line_text.push_str(&cursline);
-            curline = &mut file_data.file_lines[file_data.cursor_y as usize];
-            file_data.cursor_x = curline.line_length-(editor_data.settings.line_endings as usize);
+            file_data.file_lines[file_data.cursor_y].line_text.pop();
+            file_data.file_lines[file_data.cursor_y].line_text.pop();
+            file_data.file_lines[file_data.cursor_y].line_text.push_str(&cursline);
+            curline = &mut file_data.file_lines[file_data.cursor_y];
+            file_data.cursor_x = curline.line_length-(editor_data.settings.line_endings);
             curline.line_length = curline.line_text.len();
             return;
         }
 
-        curline.line_text.remove((file_data.cursor_x-1) as usize);
+        curline.line_text.remove((file_data.cursor_x-1));
 
         file_data.cursor_x -= 1;
